+++ CHANGELOG.md
@@ -13,13 +13,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   probably modify every file in the bundle, but should have no effect semantically.
 
 ## [1.1.0]
+
 Improves support for CommonJS projects
 
 ### Added
+
 - Changelog
 - Example invocation and usage notes for CJS projects in README
 
 ### Changed
+
 - `es-migrate` now loads .mts and .mjs migration files in addition to .ts and .js
 - `es-migrate create` generates .mts migrations instead of .ts