+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Breaking
+
+- Dropped support for Node 18
+
+### Added
+
+- Compatibility notes for less recent Node versions
+
 ### Changed
 
 - An auto-formatter has been applied to all source code. This will