+++ CHANGELOG.md
@@ -44,6 +44,6 @@ Improves support for CommonJS projects
 - Contact names and urls
 - Generated migration scripts and example migration scripts use the new package name
 
-[unreleased]: https://github.com/direct-democracy-solutions/elasticsearch-migration/compare/1.1.0...master
 [1.1.0]: https://github.com/direct-democracy-solutions/elasticsearch-migration/compare/1.0.1...1.1.0
 [1.0.1]: https://github.com/direct-democracy-solutions/elasticsearch-migration/releases/tag/1.0.1
+[unreleased]: https://github.com/direct-democracy-solutions/elasticsearch-migration/compare/1.1.0...HEAD